+++ src/main/java/lotto/domain/Lotto.java
@@ -3,7 +3,7 @@ package lotto.domain;
 import java.util.HashSet;
 import java.util.List;
 import java.util.Set;
-import lotto.constant.ExceptionMessage;
+import lotto.constant.message.ExceptionMessage;
 import lotto.constant.LottoConstant;
 import lotto.constant.RegexPattern;
