+++ src/main/java/lotto/domain/PurchaseAmount.java
@@ -1,6 +1,6 @@
 package lotto.domain;
 
-import lotto.constant.ExceptionMessage;
+import lotto.constant.message.ExceptionMessage;
 import lotto.constant.LottoConstant;
 
 public class PurchaseAmount {
