+++ src/main/java/lotto/io/input/InputValidator.java
@@ -1,6 +1,6 @@
-package lotto.io;
+package lotto.io.input;
 
-import lotto.constant.ExceptionMessage;
+import lotto.constant.message.ExceptionMessage;
 import lotto.constant.LottoConstant;
 import lotto.constant.RegexPattern;
